requirements.txt
pyalsaaudio
python-mpd2
